Update bouncing_balls.cpp
A child plays with a ball on the nth floor of a big building. The height of this floor is known. 
He lets out the ball. The ball bounces for example to two-thirds of its height. His mother looks out 
of a window that is 1.5 meters from the ground. How many times will the mother see the ball either 
falling or bouncing in front of the window?